backend/requirements.txt
# Optional but recommended for multi-encoding & i18n support
chardet>=5.2.0
charset-normalizer>=3.3.2
langdetect>=1.0.9
ftfy>=6.3.1
# ORFEAS Studio Backend Requirements
# Python dependencies for automated 3D model generation

# Core Flask and Web Framework
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6

# WebSocket support
python-socketio==5.8.0
eventlet==0.33.3

# File upload and security
werkzeug==2.3.7

# Image processing and computer vision
pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
scikit-image==0.21.0

# Machine Learning and AI
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
diffusers==0.21.4
transformers==4.35.2
accelerate==0.24.1

# 3D Processing and Mesh Generation
trimesh==3.23.5
open3d==0.17.0
numpy-stl==3.0.1
pywavefront==1.3.3
pymeshlab==2022.2.post4

# GPU and System Monitoring
gputil==1.4.0
psutil==5.9.6
nvidia-ml-py3==7.352.0

# Data handling and utilities
requests==2.31.0
aiohttp==3.8.6
asyncio-mqtt==0.13.0
websockets==11.0.3

# Configuration and environment
python-dotenv==1.0.0
pyyaml==6.0.1
configparser==6.0.0

# Development and debugging
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.9.1
flake8==6.1.0

# Monitoring and Metrics (ORFEAS PHASE 3)
prometheus-client==0.19.0      # Prometheus metrics exporter
pynvml==11.5.0                 # NVIDIA GPU monitoring (optional)

# Optional: Stable Diffusion specific (uncomment if using)
# xformers==0.0.22
# safetensors==0.4.0
# controlnet-aux==0.4.0

# Optional: Hunyuan3D integration (add if available)
# Add Hunyuan3D specific requirements here when integrating
